feature EPL-1478: adding types.ts and refactorazing mockConstants.ts to mock.ts
