+++ packages/ComponentLibrary/src/components/Modal/types.ts
@@ -1,0 +1,7 @@
+export interface ModalIProps {
+    height?: string | number;
+    width?: string | number;
+    posX?: string | number;
+    posY?: string | number;
+    children: React.ReactNode;
+  }
