+++ packages/MainUI/src/screens/Home/index.tsx
@@ -10,7 +10,7 @@ import {
   Button,
   Grid,
 } from '@workspaceui/componentlibrary/src/components//MUI';
-import { MENU_ITEMS } from '@workspaceui/componentlibrary/src/components/Modal/mockConstants';
+import { MENU_ITEMS } from '@workspaceui/componentlibrary/src/components/Modal/mock';
 import List from '@mui/material/List';
 import MenuItem from '@mui/material/MenuItem';
